requirements.txt
auth0-python==3.23.0
python-dotenv==0.20.0